refactor: CoreData 분리, 주입식으로 변경
