+++ BookApp/Data/Repository/CartCoreDataRepository.swift
@@ -1,40 +1,25 @@
 //
-//  BookStorageManager.swift
+//  CartCoreDataRepository.swift
 //  BookApp
 //
-//  Created by 정근호 on 5/12/25.
+//  Created by 정근호 on 5/16/25.
 //
 
-import CoreData
 import Foundation
+import CoreData
 
-final class CoreDataManager {
-    static let shared = CoreDataManager()
-    private init() {}
+final class CartCoreDataRepository {
+    private let context: NSManagedObjectContext
     
-    lazy var persistentContainer: NSPersistentContainer = {
-        let container = NSPersistentContainer(name: "BookApp")
-        container.loadPersistentStores(completionHandler: { (storeDescription, error) in
-            if let error = error as NSError? {
-                fatalError("Unresolved error \(error), \(error.userInfo)")
-            }
-        })
-        return container
-    }()
-    
-    var context: NSManagedObjectContext {
-        return persistentContainer.viewContext
+    init(context: NSManagedObjectContext) {
+        self.context = context
     }
     
-}
-
-// MARK: - CartStorageManager
-extension CoreDataManager: CartStorageManager {
     // MARK: - Create
     /// BookEntity가 있으면 바로 CartInfoEntity에 저장
     /// 없으면 새로 생성 후 CartInfoEntity에 저장
     func saveOrUpdateBookToCart(book: Book, quantity: Int = 1) {
-                
+        
         // BookEntity를 찾기
         let bookFetchRequest: NSFetchRequest<BookEntity> = BookEntity.fetchRequest()
         bookFetchRequest.predicate = NSPredicate(format: "isbn == %@", book.isbn)
@@ -115,7 +100,7 @@ extension CoreDataManager: CartStorageManager {
                     thumbnailURL: bookEntity.thumbnailURL ?? "",
                     quantity: Int(cartItemEntity.quantity),
                     addedDate: cartItemEntity.addedDate ?? Date()
-                    )
+                )
                 
                 let cartItem = CartItemTranslator.toDomain(dto: cartItemDTO)
                 
@@ -154,7 +139,7 @@ extension CoreDataManager: CartStorageManager {
     
     /// 해당 CartItem 장바구니에서 제거
     func removeItem(item: CartItem) {
-
+        
         guard let cartItemEntityToDelete = findCartInfoEntity(forBookISBN: item.isbn) else {
             print("삭제할 장바구니 아이템을 찾지 못했습니다 (ISBN: \(item.isbn)).")
             return
@@ -174,7 +159,7 @@ extension CoreDataManager: CartStorageManager {
     // MARK: - Update
     /// 해당 CartItem 수량 +1
     func plusQuantity(item: CartItem) {
-
+        
         guard let cartItemEntity = findCartInfoEntity(forBookISBN: item.isbn) else {
             print("수량 증가 대상 장바구니 아이템을 찾지 못했습니다 (ISBN: \(item.isbn)).")
             return
@@ -244,112 +229,5 @@ extension CoreDataManager: CartStorageManager {
         }
         
         return cartInfoList.first
-    }
-}
-
-// MARK: - RecentBookStorageManager
-extension CoreDataManager: RecentBookStorageManager {
-    
-    // MARK: - Read
-    /// 최근 본 책 정보들 Book으로 변환하여 불러옴
-    func fetchRecentBook() -> [Book] {
-        
-        let fetchRequest: NSFetchRequest<RecentBookEntity> = RecentBookEntity.fetchRequest()
-        
-        do {
-            let recentBookEntities = try context.fetch(fetchRequest)
-            
-            return recentBookEntities.map { recentBookEntity -> Book in
-                
-                return Book(
-                    authors: recentBookEntity.authors?.components(separatedBy: ", ") ?? [],
-                    contents: recentBookEntity.contents ?? "",
-                    price: Int(recentBookEntity.price),
-                    title: recentBookEntity.title ?? "",
-                    thumbnail: recentBookEntity.thumbnail ?? "",
-                    isbn: recentBookEntity.isbn ?? "")
-            }
-        } catch {
-            print("장바구니 아이템 가져오기 실패: \(error)")
-            return []
-        }
-    }
-    
-    // MARK: - Update
-    /// 최근 본 책 설정(추가), 10개 일 시 가장 먼저 추가됐던 책 제거
-    func configureRecentBook(book: Book) {
-        
-        if recentBookSize() < 10 {
-            addRecentBook(book: book)
-        } else {
-            deleteFirstRecentBook()
-            addRecentBook(book: book)
-        }
-        
-    }
-    
-    // 최근 본 책 개수
-    private func recentBookSize() -> Int {
-        let recentBookSize = fetchRecentBook().count
-        return recentBookSize
-    }
-    
-    // MARK: - Create
-    // 최근 본 책 추가
-    private func addRecentBook(book: Book) {
-        guard let recentBookEntity = NSEntityDescription.entity(forEntityName: "RecentBookEntity", in: persistentContainer.viewContext) else {
-            return
-        }
-        
-        deleteDuplicatedRecentBook(isbn: book.isbn)
-                
-        do {
-
-            let managedObject = NSManagedObject(entity: recentBookEntity, insertInto: context)
-            managedObject.setValue(book.title, forKey: "title")
-            managedObject.setValue(book.thumbnail, forKey: "thumbnail")
-            managedObject.setValue(book.price, forKey: "price")
-            managedObject.setValue(book.authors.joined(separator: ", "), forKey: "authors")
-            managedObject.setValue(book.isbn, forKey: "isbn")
-            managedObject.setValue(book.contents, forKey: "contents")
-        
-            try context.save()
-        } catch {
-            print("최근 책 추가 실패: \(error)")
-        }
-    }
-    
-    // MARK: - Delete
-    // 가장 먼저 추가됐던 최근 본 책 제거
-    private func deleteFirstRecentBook() {
-        let fetchRequest: NSFetchRequest<RecentBookEntity> = RecentBookEntity.fetchRequest()
-
-        do {
-            let recentBookEntities = try context.fetch(fetchRequest)
-            
-            let firstRecentBookEntity = recentBookEntities.first!
-            
-            context.delete(firstRecentBookEntity)
-            
-            try context.save()
-        } catch {
-            print("첫번째 삭제 실패: \(error)")
-        }
-    }
-        
-    private func deleteDuplicatedRecentBook(isbn: String) {
-        let fetchRequest: NSFetchRequest<RecentBookEntity> = RecentBookEntity.fetchRequest()
-        fetchRequest.predicate = NSPredicate(format: "isbn == %@", isbn)
-        
-        do {
-            let results = try context.fetch(fetchRequest)
-            if !results.isEmpty {
-                for objectToDelete in results {
-                    context.delete(objectToDelete)
-                }
-            }
-        } catch {
-            print("중복 삭제 실패: \(error)")
-        }
     }
 }
